src/kafka_message_handler/src/requirements.txt
kafka-python==2.2.11
pymongo==4.13.2

